feat: Переход plasma-tokens на общие механизмы
