+++ packages/plasma-tokens/package.json
@@ -25,6 +25,7 @@
     "files": ["lib", "typo", "themes", "themesValues", "theme.json", "theme@2x.json"],
     "devDependencies": {
         "@emotion/serialize": "0.11.16",
+        "@sberdevices/plasma-tokens-utils": "0.2.0",
         "@theme-ui/css": "0.3.1",
         "@types/color": "3.0.1",
         "@types/node": "^14.11.1",
